Adds a deco page worker
