+++ apps/api/src/middlewares/actors.ts
@@ -1,13 +1,12 @@
 import { RuntimeClass } from "@deco/actors";
 import { withActors } from "@deco/actors/hono";
-import { Hosts } from "@deco/sdk/hosts";
 import type { Handler } from "hono";
 import { endTime, startTime } from "hono/timing";
 import { AppEnv } from "../utils/context.ts";
 
 export const runtime = new RuntimeClass();
 
-const actorsRoutePath = `/${Hosts.API}/actors`;
+const actorsRoutePath = `/actors`;
 const actorsRoute = withActors(runtime, actorsRoutePath);
 
 export const withActorsMiddleware: Handler<AppEnv> = async (ctx, next) => {
