+++ apps/api/wrangler.toml
@@ -3,10 +3,7 @@ compatibility_flags = ["nodejs_compat"]
 name = "deco-chat-api"
 main = "main.ts"
 compatibility_date = "2024-11-27"
-routes = [
-  { pattern = "api.deco.chat", custom_domain = true },
-  { pattern = "*/*", zone_name = "deco.page" },
-]
+routes = [{ pattern = "api.deco.chat", custom_domain = true }]
 
 [vars]
 VITE_USE_LOCAL_BACKEND = false
