+++ apps/pages/README.md
@@ -1,0 +1,3 @@
+# Outbound Worker for Apps hosted on workers for platforms
+
+See initial work at: <https://github.com/deco-cx/chat/pull/496>
