+++ apps/pages/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@deco/pages",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "deno serve -A --port 3001 --unstable-hmr --unstable-kv --env-file=../web/.env main.ts",
+    "dev:wrangler": "wrangler dev --port 3001"
+  },
+  "dependencies": {
+  },
+  "devDependencies": {
+    "wrangler": "^4.13.2"
+  },
+  "engines": {
+    "node": ">=20.0.0"
+  }
+}
