+++ apps/pages/wrangler.toml
@@ -1,0 +1,17 @@
+#:schema node_modules/wrangler/config-schema.json
+compatibility_flags = ["nodejs_compat"]
+name = "deco-pages"
+main = "main.ts"
+compatibility_date = "2024-11-27"
+routes = [{ pattern = "*/*", zone_name = "deco.page" }]
+
+[observability]
+enabled = true
+
+[[dispatch_namespaces]]
+binding = "PROD_DISPATCHER"
+namespace = "deco-chat-prod"
+outbound = { service = "apps-outbound", parameters = ["DECO_CHAT_APP_ORIGIN"] }
+
+[version_metadata]
+binding = "CF_VERSION_METADATA"
